add some more VM exports for extensions
