+++ src/util/cast.js
@@ -1,5 +1,6 @@
 const Color = require('./color');
 const Sanitizer = require('./value-sanitizer');
+const isNotActuallyZero = require('./is-not-actually-zero');
 
 /**
  * @fileoverview
@@ -12,33 +13,7 @@ const Sanitizer = require('./value-sanitizer');
  * Use when coercing a value before computation.
  */
 
-/**
- * Used internally by compare()
- * @param {*} val A value that evaluates to 0 in JS string-to-number conversation such as empty string, 0, or tab.
- * @returns {boolean} True if the value should not be treated as the number zero.
- */
-const isNotActuallyZero = val => {
-    if (typeof val !== 'string') return false;
-    for (let i = 0; i < val.length; i++) {
-        const code = val.charCodeAt(i);
-        // '0'.charCodeAt(0) === 48
-        // '\t'.charCodeAt(0) === 9
-        // We include tab for compatibility with scratch-www's broken trim() polyfill.
-        // https://github.com/TurboWarp/scratch-vm/issues/115
-        // https://scratch.mit.edu/projects/788261699/
-        if (code === 48 || code === 9) {
-            return false;
-        }
-    }
-    return true;
-};
-
 class Cast {
-    /**
-     * TW: this is for extensions, it is done like this to keep parity with the old code.
-     */
-    static exports = {isNotActuallyZero, Sanitizer};
-
     /**
      * Scratch cast to number.
      * Treats NaN as 0.
